move common functionality out into util
